Tarea #3505 - agregar getUrl() y mover clase
